+++ src/ui/ControlArray.tsx
@@ -4,6 +4,7 @@ import { observable } from 'mobx';
 import { observer, Observer } from 'mobx-react';
 import { Form } from '../Form';
 import { Field } from '../Field'
+import ProxyFieldArray from '../ProxyFieldArray';
 
 import { fieldDefinition, normalizesdFieldDefinition, normalizedFormSchema } from '../interface'
 import { FieldArray } from "../FieldArray";
@@ -12,12 +13,12 @@ import { omit } from "../utils";
 interface ControlArrayProps {
 	name: string;
 	component: React.Component<any, any> | React.SFC<any> | string;
-	rules?: string;
 }
 
 @observer
 export class ControlArray extends React.Component<ControlArrayProps, any> {
-	name : string;
+	proxiedFieldsProp: ProxyFieldArray;
+	name: string;
 	field: FieldArray;
 	form : Form;
 
@@ -33,11 +34,7 @@ export class ControlArray extends React.Component<ControlArrayProps, any> {
 		_ReactiveMobxFormFieldNamePrefix: PropTypes.string.isRequired,
 	}
 
-	static defaultProps = {
-		rules: 'array'
-	}
-
-	@observable fieldsProxy: Array<any> = [];
+	@observable fieldsProp: Array<string> = [];
 
 	constructor(props, context) {
 		super(props, context);
@@ -55,19 +52,25 @@ export class ControlArray extends React.Component<ControlArrayProps, any> {
 	}
 
 	componentWillMount() {
-		// [initilaValue, rules]
-		const fieldDefinition: normalizesdFieldDefinition = this.form.formSchema[this.name] ? 
-				Field.normalizeFieldDefinition(this.form.formSchema[this.name]) : // normalize field definition from initial form schema
-				[ '', this.props.rules]; // construct normalised field definition
-
 		this.field = new FieldArray(this.name);
 		this.form.registerField(this.field);
+		this.proxiedFieldsProp = new ProxyFieldArray(this.fieldsProp, this.field.subFields);
 	}
 
 	componentWillUnmount() {
 		if (!this.field.autoRemove) {
 			this.field.setAutoRemove();
 			this.form.removeField(this.name);
+		}
+	}
+
+	componentWillReceiveProps(nextProps: ControlArrayProps, nextContext: any) {
+		const name = nextContext._ReactiveMobxFormFieldNamePrefix ? `${nextContext._ReactiveMobxFormFieldNamePrefix}.${nextProps.name}` : nextProps.name.toString();
+
+		if (this.name !== name) {
+			this.field.update(name);
+			this.name = name;
+			this.setState({}); // component Rerender -> gerChildContext -> contextUpdate
 		}
 	}
 
@@ -82,16 +85,10 @@ export class ControlArray extends React.Component<ControlArrayProps, any> {
 	render() {
 		const propsToPass = omit(this.props, ControlArray.propNamesToOmitWhenByPass);
 
-		const length =  this.fieldsProxy.length; // todo: why we need this to rerender?
+		const length =  this.fieldsProp.length; // todo: why we need this to rerender?
 
 		return React.createElement((this.props.component as any), 
-				Object.assign(
-						{}, 
-						{
-							fields: this.fieldsProxy,
-						},
-						propsToPass
-					)
+				Object.assign({}, { fields: this.proxiedFieldsProp }, propsToPass)
 			);
 	}
 }